+++ src/turnkeyml/analyze/script.py
@@ -518,7 +518,8 @@ def explore_invocation(
 
         _store_traceback(invocation_info)
 
-        raise e
+        if os.environ.get("TURNKEY_DEBUG"):
+            raise e
 
     finally:
         # Ensure that stdout/stderr is not being forwarded before updating status
